ci: break down to multiple stages \w prallel jobs
Biggest benefit of using separate stage for tests and release
is that we are not waiting for osx VM instance instance any
more on every PR, just to run
`if [[ -z "$TRAVIS_TAG" ]]; then echo "Skipping this build for non-tag builds."; exit 0; fi`
there.

It alos adds clarity to CI structure => is easier to maintain.

Signed-off-by: Alexander Bezzubov <bzz@apache.org>