ci: disable cgo by default
With go1.11 `go test` in GOPATH mode somehow
seems to depend on GCC. See https://github.com/golang/go/issues/28065

This change only enables cgo for CI profiles that
need it. Those are the ones that seem to fail
on TravisCI now, presumably due to some compiler
version missmatch.

That is a workaround and does not happen in GO11MODULE mode.

Signed-off-by: Alexander Bezzubov <bzz@apache.org>