Fix IsVendor() regex generation
The regex generation introduced by
https://github.com/go-enry/go-enry/commit/20726a1de3cdff1d64bf0a6e5f6e2475a2c96913
had a subtle but important bug in it. In particular the ^ and ^|/
prefixes were only applied to the first regex in the OR'd groupings and
not every regex in the grouping.

We fix this by wrapping each of those OR'd groups in a new non capturing
group to ensure the prefix applies to every entry in the OR'd listing.

Additionally we add three new quick tests of IsVendor to guard against
this regression in the future.

This fixes https://github.com/go-enry/go-enry/issues/135